requirements.txt
Flask==2.2.2
RPi.GPIO==0.7.1
pip==21.1.2
